Huge code cleanup regarding type casting in the whole Tulip codebase:
  - replace old style C casts by C++ dedicated ones in order to improve code readbility and explicit cast semantics + remove unneeded ones
    for base types (int, float, ...) use syntax "T(x)" instead of "(T) x"
    for object and pointer types use C++ dedicated casting operators

  - activate -Wold-style-cast warning on the whole Tulip codebase (excepted thirdparty ones) to discourage the use of old C casts

  - use static_cast instead of reinterpret_cast whenever it is possible

  - use static_cast instead of dynamic_cast whenever it is possible

+++ library/tulip-core/src/GraphMeasure.cpp
@@ -399,7 +399,7 @@ void tlp::degree(const Graph *graph, tlp::NodeStaticProperty<double> &deg,
       double normalization = 1.0;
 
       if (nbNodes > 1 && graph->numberOfEdges())
-        normalization = 1./(double) (nbNodes - 1);
+        normalization = 1./(nbNodes - 1);
 
       switch(direction) {
       case UNDIRECTED:
