+++ plugins/import/SocialNetwork/GuillaumeLatapyModel.cpp
@@ -85,7 +85,7 @@ struct GuillaumeLatapyModel:public ImportModule {
     vector<BottomNode>  vec_bottom_nodes(nbNodes);
     vector<TopNode>   vec_top_nodes(nbNodes);
 
-    unsigned int nbNodesSmallWorld = (unsigned int) ceil(0.8 * nbNodes);
+    unsigned int nbNodesSmallWorld = uint(ceil(0.8 * nbNodes));
     unsigned int nbNodesScaleFree = nbNodes - nbNodesSmallWorld;
     unsigned int maxDegreeSmallWorldNodes = 2;
     unsigned int numberOfEdges = 0;
@@ -102,7 +102,7 @@ struct GuillaumeLatapyModel:public ImportModule {
 
       if (i<nbNodesScaleFree)
         vec_bottom_nodes[i].degree =
-          (unsigned int) ceil((((((double)nbNodes)/2-10)/nbNodesScaleFree/2)*(i+1)));
+          uint(ceil(((nbNodes/2.0-10.0)/nbNodesScaleFree/2)*(i+1)));
       else
         vec_bottom_nodes[i].degree = maxDegreeSmallWorldNodes;
 
@@ -112,7 +112,7 @@ struct GuillaumeLatapyModel:public ImportModule {
 
     unsigned int degreeTop = numberOfEdges/nbNodes;
     unsigned int dixieme =
-      ((unsigned int) ceil((10.0 * numberOfEdges)/nbNodes))%10;
+      uint(ceil((10.0 * numberOfEdges)/nbNodes))%10;
 
     for (i=0; i<nbNodes-1; ++i) {
       if(i%10 >= dixieme)
